+++ app/src/androidTest/java/com/alphanication/rebirth/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.alphanication.quotexml
+package com.alphanication.rebirth
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
