+++ app/src/main/java/com/alphanication/rebirth/App.kt
@@ -1,4 +1,4 @@
-package com.alphanication.quotexml
+package com.alphanication.rebirth
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
