+++ app/src/main/java/com/alphanication/rebirth/ui/MainActivity.kt
@@ -1,8 +1,8 @@
-package com.alphanication.quotexml.ui
+package com.alphanication.rebirth.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.alphanication.quotexml.R
+import com.alphanication.rebirth.R
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
