+++ app/src/test/java/com/alphanication/rebirth/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.alphanication.quotexml
+package com.alphanication.rebirth
 
 import org.junit.Test
